+++ src/pywinrt/file_writers.h
@@ -250,7 +250,6 @@ namespace pywinrt
         w.write("import sys\n");
         w.write("import types\n");
         w.write("import typing\n");
-        w.write("import uuid\n"); // TODO: uuid is not always used
         w.write("\n");
         w.write("import @._winrt as _winrt\n", settings.module);