+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 ### Fixed
 - Fixed syntax error in type hints for map with generic value type.
 - Fixed type hints for GUIDs.
+- Fixed type hints for methods overloaded from required interfaces.
 
 ## [v1.0.0-beta.5] - 2022-06-07